Add back button to fabric
